Added down.html Added up.html Added getResults Added index2.html.bak Added results.js
[ getResults ]
The Scripts analyze the "report2.html" file that is being created by the
JSystem and create the new Format. The scripts generates a new file
called "output3.html".

[ index2.html ]
Same as index.html which contains Frame. Instead of using "report2.html"
it uses the new format (output3.html).

[ up.html,down.html ]
Used as templates to build the "outpu3.html" file.

[ results.js ]
Contains all JavaScript functions to navigate, Filter and Display in the
"output3.html" file.